refactor(arel): move visitor errors into errors.ts under ArelError
Per review on PR #993:

- Move `UnsupportedVisitError` and `NotImplementedError` into
  `packages/arel/src/errors.ts` next to the existing `ArelError` /
  `EmptyJoinError` / `BindError`. Both now extend `ArelError` so
  every arel-thrown exception shares a single `instanceof` root.

- Importers updated to pull directly from `../errors.js`; no
  re-exports through visitor.ts / to-sql.ts.

- ToSql dispatch registration now asserts at runtime that each
  registered method exists on `ToSql.prototype`. TS `keyof` can't
  constrain protected members at compile time, but the runtime guard
  catches typos/renames the moment the module imports.
