+++ packages/arel/src/visitors/index.ts
@@ -1,5 +1,5 @@
 export { ToSql } from "./to-sql.js";
-export { UnsupportedVisitError, NotImplementedError } from "./to-sql.js";
+export { UnsupportedVisitError, NotImplementedError } from "../errors.js";
 export { MySQL } from "./mysql.js";
 export { PostgreSQL, PostgreSQLWithBinds } from "./postgresql.js";
 export { SQLite } from "./sqlite.js";
